:zap: Use originally decoded data
Before, the json was being decoded but not used (to confirm it was the right format). Then it was saved, loaded from where it was saved, and decoded again.

Now, the decoded data from step one is used and then the data is saved in the background. (on the first load)